pyproject.toml
[build-system]
requires = ["setuptools>=68", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "artificial-mind-agent"
version = "0.2.0"
description = "Self-evolving multi-agent collective with shared mind-bus, layered memory, skills, generations, and discovery engine."
readme = "README.md"
license = { text = "MIT" }
requires-python = ">=3.10"
authors = [{ name = "Podtverzhdeno" }]
dependencies = [
    "python-dotenv>=1.0.0",
    "requests>=2.31.0",
]

[project.optional-dependencies]
gigachat = ["gigachat>=0.1.0"]
mcp = ["mcp>=1.0.0"]
dev = [
    "pytest>=8.0.0",
    "pytest-cov>=5.0.0",
    "ruff>=0.5.0",
]

[project.scripts]
artificial-mind = "artificial_mind.cli:main"

[tool.setuptools.packages.find]
where = ["."]
include = ["artificial_mind*"]
exclude = ["tests*", "memory*", "experiments*"]

[tool.ruff]
line-length = 100
target-version = "py310"
extend-exclude = ["memory/**", "experiments/**"]

[tool.ruff.lint]
select = ["E", "F", "W", "I", "B", "UP", "C4", "SIM"]
ignore = ["E501", "B008"]

[tool.pytest.ini_options]
testpaths = ["tests"]
addopts = "-q"
